+++ src/Global.styles.ts
@@ -1,11 +1,12 @@
 import { createGlobalStyle } from 'styled-components';
 
 const GlobalStyle = createGlobalStyle`
+  @import url('https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100;300;400;500;700;900&display=swap');
+
   html, body, div, span,
   h1, h2, h3, h4, h5, h6, p,
   a, img,b, u, i, ol, ul, li, form, label, input,
-  article, canvas, figure, figcaption, footer, header, nav{
-    @import url('https://fonts.googleapis.com/css2?family=Noto+Sans+KR:wght@100;300;400;500;700;900&display=swap');
+  article, canvas, figure, figcaption, footer, header, nav {  
     margin: 0;
     padding: 0;
     border: 0;
