+++ src/hooks/useStation.ts
@@ -1,31 +1,31 @@
-import { useState } from 'react';
+import { StationForm } from './../types';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
-import { QUERY } from '../constants/API';
-import { REGEX } from '../constants/validate';
-import { requestLines } from '../service/line';
+import useLogin from './useLogin';
 import {
   requestAddStation,
   requestDeleteStation,
   requestStations,
 } from '../service/station';
-import { Line, StationForm, StationId } from '../types';
-import useLogin from './useLogin';
+import { requestLines } from '../service/line';
+import { Line, StationId } from '../types';
+import { QUERY } from '../constants/API';
 
 const useStation = () => {
-  const [form, setForm] = useState<StationForm>({ name: '' });
-  const { name } = form;
   const { accessToken } = useLogin();
   const queryClient = useQueryClient();
   const lines = useQuery('requestLines', () => requestLines(accessToken));
 
-  const addMutation = useMutation(() => requestAddStation(form, accessToken), {
-    onSuccess: () => {
-      queryClient.invalidateQueries(QUERY.REQUEST_STATIONS);
-    },
-    onError: () => {
-      alert('역을 추가하지 못했습니다!');
-    },
-  });
+  const addMutation = useMutation(
+    (form: StationForm) => requestAddStation(form, accessToken),
+    {
+      onSuccess: () => {
+        queryClient.invalidateQueries(QUERY.REQUEST_STATIONS);
+      },
+      onError: () => {
+        alert('역을 추가하지 못했습니다!');
+      },
+    }
+  );
 
   const deleteMutation = useMutation(
     (stationId: StationId) => requestDeleteStation(stationId, accessToken),
@@ -43,14 +43,8 @@ const useStation = () => {
     requestStations(accessToken)
   );
 
-  const isValidName = REGEX.STATION_NAME.test(name);
-
-  const setName = (name: string) => {
-    setForm({ ...form, name });
-  };
-
-  const addStation = () => {
-    addMutation.mutate();
+  const addStation = (form: StationForm) => {
+    addMutation.mutate(form);
   };
 
   const isStationInLine = (lines: Line[], stationId: number) =>
@@ -71,13 +65,10 @@ const useStation = () => {
 
   return {
     stations,
-    name,
-    setName,
-    addStation,
     addMutation,
     deleteMutation,
+    addStation,
     deleteStation,
-    isValidName,
   };
 };
